requirements.txt
anyio==4.2.0
beautifulsoup4==4.12.2
black==23.11.0
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cryptography==41.0.7
fastapi==0.103.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
injector==0.21.0
Jinja2==3.1.2
libretranslatepy==2.1.1
loguru==0.7.2
lxml==5.1.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.10.3
pip==24.0
platformdirs==3.10.0
pluggy==1.4.0
pycparser==2.21
pydantic==1.10.12
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==8.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
translate==3.6.1
typing_extensions==4.9.0
Unidecode==1.2.0
urllib3==1.26.18
uvicorn==0.20.0
wheel==0.41.2
